backend/requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Caching==2.1.0

# Database
psycopg2-binary>=2.9.11

# Validation
marshmallow==3.20.1

# External APIs
anthropic==0.39.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
APScheduler==3.10.4
Werkzeug==3.0.1

# Production server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
